+++ toupy/io/__init__.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # local packages
+from .readers import *
 from .filesrw import *
 from .dataio import *
 from .h5chunk_shape_3D import *